Added support for ContentBlocks to Category
added `,itemCls: 'contentblocks_replacement'` to the definition of the contentField in line 44. This change doesn't have any effects as long as someone does not have [ContentBlocks] (https://www.modmore.com/extras/contentblocks/) installed and added msCategory to contentblocks.accepted_resource_types. But if this is the case, minishop2 will now show ContentBlocks.